+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## v1.4.0
 
 - feat(pkg:concurrent): added utility class `SlideWindowCounter`
+- feat(pkg:concurrent): added utility class `CountingRateLimiter`
 - feat(pkg:test): added utility method `asyncThrows`
 - feat(pkg:ts-types): added helper type `IJsonSafeValue`
 - feat(pkg:ts-types): added helper type `IMaybeArray`